refactor mod `utils`
